# Speed up CrystalField python tests