Changes the python `CrystalFieldTest` which uses `Fit` (or `EstimateParameters`) to reduce `MaxIterations` (or `NSamples`) used (with relaxed `assert` on the fitted parameters) so the tests runs faster. On my RHEL7 quad-core desktop the test used to take 18s, now it takes 5s. On the build servers the test now takes about 18s compared to around 30s before.

In the original issue #19471, it was suggested that these tests could all be moved into `SystemsTest` instead, which might be worthwhile for some of the longer running tests. However, these tests are not really workflow tests, and in all cases test a single fitting call (albeit one which can take quite long because they involve multiple datasets and/or large number of calls to the underlying fit functions).

**To test:**

Check that everything compiles and all tests still pass. Check that the time to run `PythonScriptsTest_CrystalFieldTest` has decreased.

Fixes #19471.

*Does not need to be in the release notes.*

---

#### Reviewer ####

Please comment on the following ([full description](http://www.mantidproject.org/Individual_Ticket_Testing)):

##### Code Review #####

- [ ] Is the code of an acceptable quality?
- [ ] Does the code conform to the [coding standards](http://www.mantidproject.org/Coding_Standards)? Is it well structured with small focussed classes/methods/functions?
- [ ] Are there unit/system tests in place? Are the unit tests small and test the a class in isolation?
- [ ] If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- [ ] Do changes function as described? Add comments below that describe the tests performed?
- [ ] How do the changes handle unexpected situations, e.g. bad input?
- [ ] Has the relevant documentation been added/updated?
- [ ] Is user-facing documentation written in a user-friendly manner?
- [ ] Has developer documentation been updated if required?

- Does everything look good? Comment with the ship it emoji but don't merge. A member of `@mantidproject/gatekeepers` will take care of it.